move db outside bot
