+++ dcoops/Dockerfile
@@ -15,4 +15,5 @@ COPY requirements.txt requirements.txt
 RUN pip3 install -r requirements.txt
 EXPOSE 80
 COPY . .
+RUN ls
 ENTRYPOINT ["python3", "-u", "./main.py"]
