+++ dcoops/bot/jigsaw.py
@@ -1,4 +1,4 @@
-import imgur
+import dcoops.bot.imgur
 from discord.ext import commands
 
 import requests
