+++ dcoops/requirements.txt
