+++ docker-compose.yml
@@ -3,7 +3,7 @@ version: '3.6'
 services: 
   python:
     build:
-      context: .
+      context: ./dcoops
       dockerfile: Dockerfile
     restart: always
     networks:
